custom eslint rule - rename
