+++ eslint-rules/index.js
@@ -1,5 +1,5 @@
 module.exports = {
   rules: {
-    'enforce-generic-array-in-union': require('./enforce-generic-array-in-union'),
+    'prefer-generic-array-type': require('./prefer-generic-array-type'),
   },
 }
