+++ package.json
@@ -76,6 +76,6 @@
     "sass": "^1.77.8",
     "typescript": "^5",
     "yaml-loader": "^0.8.1",
-    "eslint-plugin-custom-rules": "file:eslint-plugin-custom-rules"
+    "eslint-plugin-custom-rules": "file:./eslint-rules"
   }
 }
